requirements.txt
flask
fpdf
python-docx
